requirements.txt
# Core dependencies
jax==0.4.23
jaxlib==0.4.23
dm-haiku==0.0.11
optax==0.1.9
flax==0.8.0
jmp==0.0.4

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and configuration
logging>=0.5.1.2
pyyaml>=6.0.1

# Development tools
black>=23.0.0
pylint>=2.17.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: CUDA support
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jaxlib==0.4.23+cuda12.cudnn89  # Only for CUDA 12 systems
